Introductory Lectures on Convex Programming
Optimization problems arise naturally in many application fields. Whatever people do, at
some point they get a craving to organize things in a best possible way. This intention,
converted in a mathematical form, turns out to be an optimization problem of certain type.
Depending on the field of interest, it could be the optimal design problem, the optimal
control problem, the optimal location problem or even the optimal diet problem. However,
the next step, consisting in finding a solution to the mathematical model, is not so trivial.
At the first glance, everything looks very simple: Many commercial optimization packages
are easily available and any user can get a “solution” to his model just by clicking on an
icon at the screen of the personal computer. The question is, how much can we trust it?